tidy up a bit and move demof filter halpers to own folder
