+++ qcodes/utils/helpers.py
@@ -10,7 +10,6 @@ from collections import Iterator, Sequence, Mapping
 from copy import deepcopy
 
 import numpy as np
-from scipy import signal
 
 _tprint_times = {}
 
@@ -423,35 +422,3 @@ def compare_dictionaries(dict_1, dict_2,
         dicts_equal = False
     return dicts_equal, dict_differences
 
-
-def filter_win(rec, cutoff, sample_rate, numtaps, axis=-1):
-    """
-    low pass filter, returns filtered signal using FIR window
-    filter
-
-    inputs:
-        record to filter
-        cutoff frequency
-        sampling rate
-        number of frequency comppnents to use in the filer
-        axis of record to apply filter along
-    """
-    nyq_rate = sample_rate / 2.
-    fir_coef = signal.firwin(numtaps, cutoff / nyq_rate)
-    filtered_rec = signal.lfilter(fir_coef, 1.0, rec, axis=axis)
-    return filtered_rec
-
-
-def filter_ls(rec, cutoff, sample_rate, numtaps, axis=-1):
-    """
-    low pass filter, returns filtered signal using FIR
-    least squared filter
-
-    inputs:
-        record to filter
-        cutoff frequency
-        sampling rate
-        number of frequency comppnents to use in the filer
-        axis of record to apply filter along
-    """
-    raise NotImplementedError
